--- tomcat/src/main/java/nextstep/jwp/conrtroller/IndexController.java
@@ -1,13 +1,0 @@
-package nextstep.jwp.conrtroller;
-
-public class IndexController {
-    private String requestMethod;
-    private String requestUrl;
-
-    public IndexController(String requestMethod, String requestUrl) {
-        this.requestMethod = requestMethod;
-        this.requestUrl = requestUrl;
-    }
-
-
-}
